Refactor out common pop-up-window view code into mixins
